move cache_seed_gen to core/utils.py
