+++ rdagent/core/utils.py
@@ -87,12 +87,35 @@ def import_class(class_path: str) -> Any:
     return getattr(module, class_name)
 
 
+class CacheSeedGen:
+    """
+    It is a global seed generator to generate a sequence of seeds.
+    This will support the feature `use_auto_chat_cache_seed_gen` claim
+
+    NOTE:
+    - This seed is specifically for the cache and is different from a regular seed.
+    - If the cache is removed, setting the same seed will not produce the same QA trace.
+    """
+
+    def __init__(self) -> None:
+        self.set_seed(RD_AGENT_SETTINGS.init_chat_cache_seed)
+
+    def set_seed(self, seed: int) -> None:
+        random.seed(seed)
+
+    def get_next_seed(self) -> int:
+        """generate next random int"""
+        return random.randint(0, 10000)
+
+cache_seed_gen = CacheSeedGen()
+
+
 def _subprocess_wrapper(f: Callable, seed: int, args: list) -> Any:
     """
     It is a function wrapper. To ensure the subprocess has a fixed start seed.
     """
-    from rdagent.oai.llm_utils import APIBackend
-    APIBackend.cache_seed_gen.set_seed(seed)
+
+    cache_seed_gen.set_seed(seed)
     return f(*args)
 
 
@@ -120,10 +143,9 @@ def multiprocessing_wrapper(func_calls: list[tuple[Callable, tuple]], n: int) ->
     if n == 1:
         return [f(*args) for f, args in func_calls]
 
-    from rdagent.oai.llm_utils import APIBackend
     with mp.Pool(processes=max(1, min(n, len(func_calls)))) as pool:
         results = [
-            pool.apply_async(_subprocess_wrapper, args=(f, APIBackend.cache_seed_gen.get_next_seed(), args))
+            pool.apply_async(_subprocess_wrapper, args=(f, cache_seed_gen.get_next_seed(), args))
             for f, args in func_calls
         ]
         return [result.get() for result in results]
